+++ CHANGELOG.md
@@ -1,3 +1,25 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+All notable changes to the "Your Package Name" package will be documented in this file.
+
+## [1.0.8] - 07/11/2023
+
+- Added a concept of a thumbnail controller, which allows users to generate or initialize thumbnails and stores the path of the thumbnail.
+- Introduced a new widget called `ThumbnailTile`, enabling users to seamlessly generate thumbnails and use them in their Flutter application.
+
+## [1.0.7] 
+
+- Added functionalities to generate video thumbnails.
+
+## [1.0.5] 
+
+- Introduced the `VideoDetails` class, responsible for storing video details such as the video name, video path, and video size.
+
+## [1.0.1] 
+
+- Lists all video files from local storage.
+
+## [0.0.1] 
+
+- Initial package generation.
+